Street prediction using nominatim reverse geocoding added
The app can now predict the street you are on
from sending an API call to nominatim's reverse
geocoder using the location of the housenumber.
It will send this API call when you open the
keypad activity.

It also only does this when internet connection
is available, otherwise the street name will be
set to whatever the streetname of the previous
address was.