reworked saving mechanism
All information is saved to app's internal storage,
and when the user decides to save, they are
presented with the documents provider from which
they can choose a location to store a zip file
containing all the osm files and image files
from the internal storage. Due to using the
documents provider, the app no longer needs
the storage permission.