added colors to street name and building levels in keypad
In the keypad, if the street name of the current
housenumber you are trying to enter is the same
as the street name of the previous address, the
color becomes green. Otherwise, it becomes red. This
is useful to figure out if the reverse geocoding algorithm
found a street that is different to the last one.

Also, the six building levels options all have
six different colors. This should help
differenciating them. There are different
sets of colors for light mode and dark mode.