files save in seperate thread
files are now zipped and saved in a seperate
thread. This prevents the document picker from
being useable while files are being saved. Instead
of this, a dialog telling the user that files
are saving shows up (until the files have saved).
This is to prevent the user from adding markers to
the map while data is being saved. (As markers
are deleted immediately after data is saved).

The reason data has to be saved in another thread
& a dialog needs to be shown is because if multiple
image notes have been taken, zipping all of them is
a quite slow process.