added menu & changed markerlist to hashmap
A menu, containing the option to save the data,
"recover" saved data and a currently unused
"upload to OSM" option has been added to the
app.

The markerlist used to store markers is
now instead a markerhashmap, where the key is
the sql ID for the marker, and the value is the
actual marker itself. This allows for accessing the
sql item of a marker if needed.

The mapView and some related functions have been
moved to a seperate class called "Map". The map
is still currently public and so is just as easily
accessible as before.