added automcomplete feature for manual street name
When you try and change the street name now, the
app automatically attempts to get all the
street names in a 100m radius using an overpass
query. This is then displayed as autocomplete
suggestions when typing in the street name.

Also, you can now modifiy the street name by
pressing on the "Street Name:" textview or by
clicking on the street_name_value textview.