Save previously added custom imageries
Previously added custom imageries are saved to
preferences. They are shown as a list of options
in the "customImageryDialog" every time that is
opened up in order to change the custom imagery.
You can click on the previously added custom imageries
to autofill the custom Imagery textbox with it.

You can also delete the previously added custom imagery
from the list by pressing the "cross" button